+++ COT/tests/test_info.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python
 #
-# info.py - Unit test cases for COTInfo class.
+# test_info.py - Unit test cases for COTInfo class.
 #
 # January 2015, Glenn F. Matthews
-# Copyright (c) 2013-2015 the COT project developers.
+# Copyright (c) 2013-2016 the COT project developers.
 # See the COPYRIGHT.txt file at the top-level directory of this distribution
 # and at https://github.com/glennmatthews/cot/blob/master/COPYRIGHT.txt.
 #
@@ -108,6 +108,7 @@ Files and Disks:     File Size  Capacity Device
                      --------- --------- --------------------
   input.vmdk         149.00 kB   1.00 GB harddisk @ SCSI 0:0
   input.iso          352.00 kB           cdrom @ IDE 1:0
+  sample_cfg.txt          78 B
 
 Hardware Variants:
   System types:             vmx-07 vmx-08
@@ -163,6 +164,7 @@ Files and Disks:     File Size  Capacity Device
                      --------- --------- --------------------
   input.vmdk         149.00 kB   1.00 GB harddisk @ SCSI 0:0
   input.iso          352.00 kB           cdrom @ IDE 1:0
+  sample_cfg.txt          78 B
 
 Hardware Variants:
   System types:             vmx-07 vmx-08
@@ -209,7 +211,12 @@ Version:  DEV
 Files and Disks:     File Size  Capacity Device
                      --------- --------- --------------------
   input.vmdk         149.00 kB   1.00 GB harddisk @ SCSI 0:0
+    File ID: file1
+    Disk ID: vmdisk1
   input.iso          352.00 kB           cdrom @ IDE 1:0
+    File ID: file2
+  sample_cfg.txt          78 B
+    File ID: textfile
 
 Hardware Variants:
   System types:             vmx-07 vmx-08
@@ -339,6 +346,8 @@ Version:  15.4(2.4)T
 Files and Disks:     File Size  Capacity Device
                      --------- --------- --------------------
   input.vmdk         149.00 kB   1.00 GB harddisk @ IDE 0:0
+    File ID: vios-adventerprisek9-m.vmdk
+    Disk ID: vios-adventerprisek9-m.vmdk
   (disk placeholder)        -- 128.00 MB harddisk @ IDE 0:1
 
 Hardware Variants:
@@ -473,6 +482,8 @@ End User License Agreement(s):
 Files and Disks:     File Size  Capacity Device
                      --------- --------- --------------------
   input.vmdk         149.00 kB   1.00 GB harddisk @ SCSI 0:0
+    File ID: file1
+    Disk ID: vmdisk1
 
 Hardware Variants:
   System types:             vmx-04
@@ -532,6 +543,8 @@ Properties:
 Files and Disks:     File Size  Capacity Device
                      --------- --------- --------------------
   input.vmdk         149.00 kB   1.00 GB harddisk @ SCSI 0:0
+    File ID: file1
+    Disk ID: vmdisk1
 
 Hardware Variants:
   System types:             vmx-08
@@ -570,6 +583,8 @@ Properties:
 Files and Disks:        File Size  Capacity Device
                         --------- --------- --------------------
   ubuntu.2.0-disk1.vmdk             8.00 GB harddisk @ SATA 0:0
+    File ID: file1
+    Disk ID: vmdisk1
 
 Hardware Variants:
   System types:             virtualbox-2.2
@@ -637,7 +652,10 @@ Version:  (No version string)
 Files and Disks:                       File Size  Capacity Device
                                        --------- --------- --------------------
   this_is_a_really_long_filename_fo... 149.00 kB   1.00 GB 
+    File ID: input.vmdk
+    Disk ID: input.vmdk
   input.iso                            352.00 kB           cdrom @ (?) ?:0
+    File ID: input.iso
   (disk placeholder)                          -- 128.00 MB 
 
 Configuration Profiles:             CPUs    Memory NICs Serials Disks/Capacity
@@ -741,7 +759,10 @@ Version:  (No version string)
 Files and Disks:     File Size  Capacity Device
                      --------- --------- --------------------
   this_is_a_reall... 149.00 kB   1.00 GB 
+    File ID: input.vmdk
+    Disk ID: input.vmdk
   input.iso          352.00 kB           cdrom @ (?) ?:0
+    File ID: input.iso
   (disk placeholder)        -- 128.00 MB 
 
 Configuration Profiles:             CPUs    Memory NICs Serials Disks/Capacity